requirements.txt
streamlit 
pyvis 
network 
pandas